Create README.md
Socket programming, also known as network programming, is a concept that involves writing programs that can communicate over a network. It is a fundamental aspect of developing applications that need to exchange data between devices, either locally or over the internet. Sockets provide a standard mechanism for processes on different devices to communicate with each other.